webmonitor: choose different query timing to get nicer graph
the web interface used to miss a lot of data, because it queried
results up to now(). Instead of that, let's ask till "up to some
delay ago" (on the order of 1s), which allows for small time differences
between data recording, network transfer, writing to database.

This makes nicer graph, and not missed data.

Could be further improved by querying on ObjectID together with date,
but that's for the future.

Also, the query repeat time is changed from 2s to 3s.